changed inc matrix to int type + fixed access to juncs numbers

+++ grammar/driver/driver.hh
@@ -34,7 +34,7 @@ private:
   OurFlexLexer *plex_; // maybe this is good name
   std::vector<std::string> lines_of_prog;
 
-  MX::Matrix<double> incidence_;
+  MX::Matrix<int> incidence_;
   MX::Matrix<double> resistance_;
   MX::Matrix<double> eds_;
